Change the method for determining when exception table entries need to be added for inlined subroutines, to properly handle exceptions whose protected areas overlap the inlined subroutine, but begin before or end after the inlined code.
Ondrej discovered the bug when he ran Soot on the
sun.nio.ch.SocketChannelImpl.read(java.nio.ByteBuffer[]) in
j2sdk1.4.2's standard library (which caused Soot to throw a
RuntimeException).

This is a partial fix, because it doesn't properly handle
protected areas which extend to the end of the method body (and
there even is such an area in the case of
SocketChannelImpl.read()).  Until we come up with a convention
for what Trap.getEndUnit() should return for traps that extend to
the end of the method, we just lie, and cut the protected range
one statement short (thanks to Ondrej for that suggestion).