Factored some graph drawing code into CFGToDotGraph.java, so it would be available for dumping graphs between phases.
Add a setGraphAttribute(DotGraphAttribute) method to some
DotGraph elements.

Have DotGraph take its output filename as an argument, instead of
basing it on the method name, to make it easier to dump DotGraphs
of successive versions of the same CFG.

Move the code for abbreviating the String representation
of ThrowableSets from CFGViewer to ThrowableSet.

Add instrumentation to report the number of registered ThrowableSets.