bugfix for this situation:
Consider those methods:

foo(Iterator i) {
 bar(i);
 baz(i);
}

bar(Iterator i) {
 i.hasNext();
}

baz(Iterator i) {
 i.next();
}

The important point is the following: We have some method "baz" which
cannot be optimized and it refers to the same iterator as another
method "bar". Now when we analyze "bar" we see that it cannot hit
final, because we only propagate bindings generated within "bar" and
at the end of "bar" we are always in the initial configuration. Hence,
we eliminate the hasNext-shadow in "bar". This is not sound.

We have to bind non-local bindings (positively)! However, we *Can* do an early abort: If a final state is hit on the synthetic final unit, there is nothing to rescue anyway. Hence abort early then.