reworked strong must-alias analysis;
previously it could happen that an instance key was invalidated although it was not contained in a loop! this could happen if its definition was behind complicated control flow, i.e. the definition could be reached over more than one branch. now we use SCCs to determine whether a variable is redefined within a loop.

also, we moved the analysis and the class InstanceKey from abc to Soot