Added new SOOT interface function getMethodDescriptor.
This is a replacement for the MethodDescriptorTag whos introduction will
be reverted in the next commit.
The getMethodDescriptor function can return NULL in cases the SootMethod
is not derived from ASM Objectweb. Otherwise it returns the binary
method descriptor. ...Also for external function which was not the case
for the old MethodDescriptorTag.