Make results reproducable.
For the TypeResolver some local names needed to be prefixed with $
to have same handled by the LocalNAmeStandardizer class which replaces
the random names through up counting numbers.

For AsmMethodSource the traps are now added to the initial convert
worklist in a deterministic manner which results in consitent stack
variable naming.