Do not split new assignments.
We want to avoid creating useless temporary variables
in the split_new method by spliting up the initial local
variable assignments. This creates overhead and more
importantly creates alias to the same object which
complicates later analysis.
This is because the way in which the split is implemented,
the <init> method is "executed" on the new introduced
tmp local variable, whereas following code works on the
original local.