Fix local variable mixup in case of no debug information.
If a local variable was pushed to stack and than reassigned, the local
variable representing the stack value got assigned the same name as the
original local which is invalid.
We solve this by running the Aggregator before the LocalSplitter
(previously the Agrregator did not work on splitted locals because it
did not handle the "#" suffix like other transforms do) and not using
the CopyPropagator at all. The CopyPropagator seems to be a less
elaborate aggregator and is now not need anymore since the aggregator is
working correctly.