Workaround CFG build for body-less methods
For some reason Soot is attempting to build a CFG for (abstract) methods
that have no body attached, since they are declared in the interfaces.
Looks like a rather werid corner case that was not taken into account.

Rather than processing such methods, we just skip them, which does not
appear to affect the overall construction in usual circumstances.