Fix race condition in AsmMethodSource.
The code we execute in parallel used global mutable state to translate
method descriptors of dynamic invoke instructions.
We now use a new instance of the translation class. Giving the number of
usually low number of dynamic invoke instructions it is not too
important to lose the caching done in the Util class.